# soul-shield/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/tasks/history": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "from/to না দিলে গত ৭ দিন default হিসেবে দেখাবে (আজসহ)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Get task history for a date range",
                "parameters": [
                    {
                        "type": "string",
                        "description": "YYYY-MM-DD",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "YYYY-MM-DD",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/task.TaskWithStatusResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/task.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tasks/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Hard delete করে, কিন্তু completion history অক্ষত থাকে (snapshot হিসেবে)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Delete task",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/task.SuccessResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/task.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/task.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Owner (personal task) অথবা Admin (global task) আপডেট করতে পারবে",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Update task",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/task.UpdateTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/task.TaskResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/task.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/task.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/task.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "task.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid request"
                }
            }
        },
...
        "task.SubTaskInput": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "update এ existing sub-task ম্যাচ করাতে; create এ omit করবেন",
                    "type": "integer",
                    "example": 5
                },
                "target_count": {
                    "description": "task_type=counter হলে required",
                    "type": "integer",
                    "example": 10
                },
                "task_type": {
                    "type": "string",
                    "enum": [
                        "normal",
                        "counter"
                    ],
                    "example": "normal"
                },
                "title": {
                    "type": "string",
                    "example": "Read 1 page"
                }
            }
        },
        "task.SubTaskStatusResponse": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "type": "string"
                },
                "progress_count": {
                    "type": "integer"
                },
                "status": {
                    "type": "string",
                    "example": "pending"
                },
                "sub_task_id": {
                    "type": "integer",
                    "example": 5
                },
                "target_count": {
                    "type": "integer",
                    "example": 10
                },
                "task_type": {
                    "type": "string",
                    "example": "normal"
                },
                "title": {
                    "type": "string",
                    "example": "Read 1 page"
                }
            }
        },
        "task.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Task created successfully"
                }
            }
        },
        "task.TaskResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer",
                    "example": 1
                },
                "description": {
                    "type": "string",
                    "example": "Read at least 1 page daily"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "is_global": {
                    "type": "boolean",
                    "example": false
                },
                "owner_id": {
                    "type": "integer",
                    "example": 3
                },
                "recurrence_days": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        1,
                        3,
                        5
                    ]
                },
                "recurrence_type": {
                    "type": "string",
                    "example": "custom"
                },
                "sub_tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/task.SubTaskStatusResponse"
                    }
                },
                "title": {
                    "type": "string",
                    "example": "Read Quran"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "task.TaskWithStatusResponse": {
            "type": "object",
            "properties": {
                "category_color": {
                    "type": "string"
                },
                "category_id": {
                    "type": "integer"
                },
                "category_name": {
                    "type": "string"
                },
                "completed_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "is_global": {
                    "type": "boolean"
                },
                "progress_count": {
                    "type": "integer"
                },
                "recurrence_days": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "recurrence_type": {
                    "type": "string"
                },
                "reminder_time": {
                    "type": "string"
                },
                "reward_text": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "sub_tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/task.SubTaskStatusResponse"
                    }
                },
                "target_count": {
                    "type": "integer"
                },
                "task_id": {
                    "type": "integer"
                },
                "task_type": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "task.UpdateTaskRequest": {
            "type": "object",
            "properties": {
                "category_id": {
                    "description": "0 পাঠালে uncategorize হবে",
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "recurrence_days": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "recurrence_type": {
                    "type": "string",
                    "enum": [
                        "daily",
                        "weekly",
                        "custom"
                    ]
                },
                "reminder_time": {
                    "description": "\"HH:MM\" 24-hour format, empty string clears it",
                    "type": "string"
                },
                "reward_text": {
                    "type": "string"
                },
                "sub_tasks": {
                    "description": "SubTasks nil হলে touch করা হবে না; দিলে (even []) পুরো লিস্ট replace হবে -\nexisting id ম্যাচ করলে update, নতুন id ছাড়া entry হলে insert, বাদ পড়া entry delete।",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/task.SubTaskInput"
                    }
                },
                "target_count": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    }
}